Split files model out the models.py file
