+++ projecto/apiv1/files/api.py
@@ -5,7 +5,7 @@ from flask.ext.login import current_user
 from kvkit import NotFoundError
 
 from ..hacks import Blueprint
-from ...models import File
+from .models import File
 from ...utils import ensure_good_request, project_access_required, jsonify
 
 blueprint = Blueprint("api_v1_files", __name__,
